explaniton
I'm going to write here what i did in each language.HTML- i created an header and a sub-header that explains the purpose of the page. I added 2 text areas that allows to the user to put his name and his feelings and I also created a button element that sends the details that the user put on his feelings to the API. CSS- in css I designed the button and the background of the page with a photo. I also made a loading design that activated every time we wait the API will return us an answer about how we feel. the loading design make the whole page blue with a white square inside of him  and the square is revolving around himself. every time the square is turning in 90 degree, the square becomes empty from color and when another turn is completed the square is filled with white, and so on.
JS- I created a variable that receives everything we got from the API. then we check if the status of the variable is above 400. if so, we will display the error to the user. if the status is lower then 400 we will proceed to the rest of the code. in the rest of the code we check the polarity that we got from the API. if the polarity is above 0, we will write in green the word "positive" and the number of the polarity. if we got less then 0 we will write "negative" and the number in red. if the polarity is 0 we will write "neutral" and the number 0 in gray. then we will add this function as an event to the button to make the function happen when we press the button.